# backend/api/fenzvideo/v1/channel_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.33.4
// source: fenzvideo/v1/channel.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationChannelServiceGetChannel = "/fenzvideo.v1.ChannelService/GetChannel"
const OperationChannelServiceSubscribe = "/fenzvideo.v1.ChannelService/Subscribe"
const OperationChannelServiceUnsubscribe = "/fenzvideo.v1.ChannelService/Unsubscribe"

type ChannelServiceHTTPServer interface {
	GetChannel(context.Context, *GetChannelRequest) (*ChannelReply, error)
	Subscribe(context.Context, *SubscribeRequest) (*MembershipReply, error)
	Unsubscribe(context.Context, *UnsubscribeRequest) (*UnsubscribeReply, error)
}

func RegisterChannelServiceHTTPServer(s *http.Server, srv ChannelServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/channels/{id}", _ChannelService_GetChannel0_HTTP_Handler(srv))
	r.POST("/api/v1/channels/{id}/subscribe", _ChannelService_Subscribe0_HTTP_Handler(srv))
	r.DELETE("/api/v1/channels/{id}/subscribe", _ChannelService_Unsubscribe0_HTTP_Handler(srv))
}

func _ChannelService_GetChannel0_HTTP_Handler(srv ChannelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetChannelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChannelServiceGetChannel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetChannel(ctx, req.(*GetChannelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ChannelReply)
		return ctx.Result(200, reply)
	}
}

func _ChannelService_Subscribe0_HTTP_Handler(srv ChannelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SubscribeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChannelServiceSubscribe)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Subscribe(ctx, req.(*SubscribeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MembershipReply)
		return ctx.Result(200, reply)
	}
}

func _ChannelService_Unsubscribe0_HTTP_Handler(srv ChannelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UnsubscribeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationChannelServiceUnsubscribe)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Unsubscribe(ctx, req.(*UnsubscribeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UnsubscribeReply)
		return ctx.Result(200, reply)
	}
}

type ChannelServiceHTTPClient interface {
	GetChannel(ctx context.Context, req *GetChannelRequest, opts ...http.CallOption) (rsp *ChannelReply, err error)
	Subscribe(ctx context.Context, req *SubscribeRequest, opts ...http.CallOption) (rsp *MembershipReply, err error)
	Unsubscribe(ctx context.Context, req *UnsubscribeRequest, opts ...http.CallOption) (rsp *UnsubscribeReply, err error)
}

type ChannelServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewChannelServiceHTTPClient(client *http.Client) ChannelServiceHTTPClient {
	return &ChannelServiceHTTPClientImpl{client}
}

func (c *ChannelServiceHTTPClientImpl) GetChannel(ctx context.Context, in *GetChannelRequest, opts ...http.CallOption) (*ChannelReply, error) {
	var out ChannelReply
	pattern := "/api/v1/channels/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChannelServiceGetChannel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChannelServiceHTTPClientImpl) Subscribe(ctx context.Context, in *SubscribeRequest, opts ...http.CallOption) (*MembershipReply, error) {
	var out MembershipReply
	pattern := "/api/v1/channels/{id}/subscribe"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationChannelServiceSubscribe))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ChannelServiceHTTPClientImpl) Unsubscribe(ctx context.Context, in *UnsubscribeRequest, opts ...http.CallOption) (*UnsubscribeReply, error) {
	var out UnsubscribeReply
	pattern := "/api/v1/channels/{id}/subscribe"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationChannelServiceUnsubscribe))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}