# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/auth/login:
        post:
            tags:
                - AuthService
            operationId: AuthService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.LoginReply'
    /api/v1/auth/refresh:
        post:
            tags:
                - AuthService
            operationId: AuthService_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.RefreshTokenReply'
    /api/v1/auth/register:
        post:
            tags:
                - AuthService
            operationId: AuthService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.RegisterReply'
...
    /api/v1/channels/{id}:
        get:
            tags:
                - ChannelService
            operationId: ChannelService_GetChannel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.ChannelReply'
    /api/v1/channels/{id}/subscribe:
        post:
            tags:
                - ChannelService
            operationId: ChannelService_Subscribe
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.SubscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.MembershipReply'
        delete:
            tags:
                - ChannelService
            operationId: ChannelService_Unsubscribe
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.UnsubscribeReply'
...
    /api/v1/tags/my:
        get:
            tags:
                - TagService
            operationId: TagService_GetMyTags
            parameters:
                - name: sessionId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.TagListReply'
        put:
            tags:
                - TagService
            operationId: TagService_SetMyTags
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/fenzvideo.v1.SetMyTagsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/fenzvideo.v1.TagListReply'
...
components:
    schemas:
...
        fenzvideo.v1.ChannelReply:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: string
                displayName:
                    type: string
                avatarUrl:
                    type: string
                monthlyFee:
                    type: number
                    format: double
                subscriberCount:
                    type: string
                membershipStatus:
                    type: string
...
        fenzvideo.v1.LoginReply:
            type: object
            properties:
                id:
                    type: string
                username:
                    type: string
                displayName:
                    type: string
                role:
                    type: string
                token:
                    type: string
                refreshToken:
                    type: string
        fenzvideo.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        fenzvideo.v1.MembershipReply:
            type: object
            properties:
                status:
                    type: string
                tier:
                    type: string
        fenzvideo.v1.RefreshTokenReply:
            type: object
            properties:
                token:
                    type: string
                refreshToken:
                    type: string
        fenzvideo.v1.RefreshTokenRequest:
            type: object
            properties:
                refreshToken:
                    type: string
        fenzvideo.v1.RegisterReply:
            type: object
            properties:
                id:
                    type: string
                username:
                    type: string
                displayName:
                    type: string
                token:
                    type: string
                refreshToken:
                    type: string
        fenzvideo.v1.RegisterRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                displayName:
                    type: string
        fenzvideo.v1.SetMyTagsRequest:
            type: object
            properties:
                tagIds:
                    type: array
                    items:
                        type: string
                sessionId:
                    type: string
        fenzvideo.v1.SubscribeRequest:
            type: object
            properties:
                id:
                    type: string
        fenzvideo.v1.TagItem:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                slug:
                    type: string
        fenzvideo.v1.TagListReply:
            type: object
            properties:
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/fenzvideo.v1.TagItem'
...
        fenzvideo.v1.UnsubscribeReply:
            type: object
            properties:
                status:
                    type: string
...